refactor(api): split disable from user service
