+++ packages/api/src/modules/user/services/rgpd/user.rgpd.service.ts
@@ -5,6 +5,9 @@ import userResetRepository from "../../repositories/user-reset.repository";
 import consumerTokenRepository from "../../repositories/consumer-token.repository";
 import { uniformizeId } from "../../../../shared/helpers/RepositoryHelper";
 import statsService from "../../../stats/stats.service";
+import notifyService from "../../../notify/notify.service";
+import { NotificationType } from "../../../notify/@types/NotificationType";
+import userRepository from "../../repositories/user.repository";
 
 export class UserRgpdService {
     public async getAllData(userId: string): Promise<UserDataDto> {
@@ -32,6 +35,27 @@ export class UserRgpdService {
                 associationVisit: associationVisits.map(userIdToString),
             },
         };
+    }
+
+    public async disable(userId: string) {
+        const user = await userService.getUserById(userId);
+        if (!user) return false;
+        // Anonymize the user when it is being deleted to keep use stats consistent
+        // It keeps roles and signupAt in place to avoid breaking any stats
+        const disabledUser = {
+            ...user,
+            active: false,
+            email: "",
+            jwt: null,
+            hashPassword: "",
+            disable: true,
+            firstName: "",
+            lastName: "",
+        };
+
+        notifyService.notify(NotificationType.USER_DELETED, { email: user.email });
+
+        return !!(await userRepository.update(disabledUser));
     }
 }
